Fix some problematic behavior around vertex buffer strides.
Vertex buffer strides were handled differently on Direct3D and other
backends. When a ShaderSet with an input layout that didn't match the
vertex buffer layout was used, all backends except for D3D would fail.
This was hidden for the most part because the shaders were authored in
such a way as to include the unnecessary vertex input elements, and also
to ignore failures to set missing vertex attributes. I've fixed up this
handling.

As part of this, I removed the previously-useless "Stride" parameter on
VertexDescriptor. This was always being set to 0, and if a stride was
used, it was read from the total size of the vertex (which is what the
stride is anyways).